Show root parameter affects node fill
Before, show root parameter was affecting the tree structure - the tree would either being from the root or the next descendant, depending on the value of show_root. Now, it always begins at the root, but the root node will not appear in the legend if "show_root" is set to false.
Parameter default changed to FALSE.